+++ src/ui/fortsettings.h
@@ -72,7 +72,7 @@ private:
     bool iniBool(const QString &key, bool defaultValue = false) const;
     int iniInt(const QString &key, int defaultValue = 0) const;
     uint iniUInt(const QString &key, int defaultValue = 0) const;
-    int iniReal(const QString &key, qreal defaultValue = 0) const;
+    qreal iniReal(const QString &key, qreal defaultValue = 0) const;
     QString iniText(const QString &key, const QString &defaultValue = QString()) const;
     QStringList iniList(const QString &key) const;
